fix: get right song
